Improve fees tests
